Add handling style.css and refactor handlers as well
Now we provide whole postgres store and gitlab client structs to handlers.Start() function, but isolate behaviour with interfaces using for each handler
