+++ internal/web/style.css
@@ -1,0 +1,12 @@
+#dev-th {
+    padding-left: 0;
+    text-align: left;
+}
+
+.project-th {
+    padding-right: 1.5em;
+}
+
+.dev-td {
+    padding-right: 1.5em;
+}
